fix: remove default owner tag
This always defaulted to dscott.  We should remove the owner requirement as well as
the default owner tag and enforce this another way if we want an owner enforced.  This
is due to the high probability that organizations that consume this module do not use
an owner tag and have other tagging requirements.

Signed-off-by: Dustin Scott <dustin.scott18@gmail.com>