fix: replicas and autoscaling.*_replicas
This commit fixes a change in the OCM API that will reject requests that include
values for both replicas and autoscaling.min_replicas or autoscaling.max_replicas
for a machine pool.  Previously, we would use the same value and this was OK, but
now it appears the OCM API rejects these requests.

Signed-off-by: Dustin Scott <dscott@redhat.com>